# docs/docs_swagger/swagger.yaml
basePath: /api
...
host: localhost:10031
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: 智能门禁管理系统API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ILock HTTP Service API
  version: "1.0"
paths:
...
  /admin/device:
    delete:
      consumes:
      - application/json
      description: 删除一个或多个设备
      parameters:
      - description: 设备ID列表
        in: body
        name: ids
        required: true
        schema:
          type: object
      - description: 设备ID数组
        in: body
        name: ids.ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功消息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除设备
      tags:
      - Device
    get:
      consumes:
      - application/json
      description: 根据查询条件获取设备列表，包含设备状态
      parameters:
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - default: 1
        description: 页码
        in: query
        name: pageNum
        type: integer
      - default: true
        description: 是否降序
        in: query
        name: desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 返回设备列表和分页信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取设备列表
      tags:
      - Device
    post:
      consumes:
      - application/json
      description: 创建一个新的设备
      parameters:
      - description: 设备信息
        in: body
        name: device
        required: true
        schema:
          type: object
      - description: 设备ID
        in: formData
        name: deviceId
        required: true
        type: string
      - description: 建筑ID
        in: formData
        name: buildingId
        required: true
        type: integer
      - description: 设备设置
        in: formData
        name: settings
        type: object
      - description: 欠费更新间隔(秒)
        in: formData
        name: settings.arrearageUpdateDuration
        type: integer
      - description: 通知更新间隔(秒)
        in: formData
        name: settings.noticeUpdateDuration
        type: integer
      - description: 广告更新间隔(秒)
        in: formData
        name: settings.advertisementUpdateDuration
        type: integer
      - description: 广告播放时长(秒)
        in: formData
        name: settings.advertisementPlayDuration
        type: integer
      - description: 通知播放时长(秒)
        in: formData
        name: settings.noticePlayDuration
        type: integer
      - description: 空闲时长(秒)
        in: formData
        name: settings.spareDuration
        type: integer
      - description: 通知停留时长(秒)
        in: formData
        name: settings.noticeStayDuration
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回创建的设备信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建设备
      tags:
      - Device
    put:
      consumes:
      - application/json
      description: 更新设备信息，包括设备ID、所属建筑和设置
      parameters:
      - description: 设备更新信息
        in: body
        name: device
        required: true
        schema:
          type: object
      - description: 设备ID
        in: formData
        name: id
        required: true
        type: integer
      - description: 设备ID
        in: formData
        name: deviceId
        type: string
      - description: 建筑ID
        in: formData
        name: buildingId
        type: integer
      - description: 设备设置
        in: formData
        name: settings
        type: object
      - description: 欠费更新间隔(秒)
        in: formData
        name: settings.arrearageUpdateDuration
        type: integer
      - description: 通知更新间隔(秒)
        in: formData
        name: settings.noticeUpdateDuration
        type: integer
      - description: 广告更新间隔(秒)
        in: formData
        name: settings.advertisementUpdateDuration
        type: integer
      - description: 广告播放时长(秒)
        in: formData
        name: settings.advertisementPlayDuration
        type: integer
      - description: 通知播放时长(秒)
        in: formData
        name: settings.noticePlayDuration
        type: integer
      - description: 空闲时长(秒)
        in: formData
        name: settings.spareDuration
        type: integer
      - description: 通知停留时长(秒)
        in: formData
        name: settings.noticeStayDuration
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回更新后的设备信息，包含状态
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新设备
      tags:
      - Device
...
  /admin/device/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取设备详细信息
      parameters:
      - description: 设备ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回设备详细信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 设备不存在
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取单个设备
      tags:
      - Device
  /admin/devices:
    post:
      consumes:
      - application/json
      description: 批量创建多个设备
      parameters:
      - description: 设备信息数组
        in: body
        name: devices
        required: true
        schema:
          type: object
      - description: 设备数组
        in: body
        name: devices.devices
        required: true
        schema:
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 返回创建的设备信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 批量创建设备
      tags:
      - Device
...
  /device/client/advertisements:
    get:
      consumes:
      - application/json
      description: 根据设备ID获取分配给该设备的广告
      produces:
      - application/json
      responses:
        "200":
          description: 返回设备广告列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 未找到设备
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: 获取设备广告
      tags:
      - Device
...
  /device/client/notices:
    get:
      consumes:
      - application/json
      description: 根据设备ID获取分配给该设备的通知
      produces:
      - application/json
      responses:
        "200":
          description: 返回设备通知列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 未找到设备
          schema:
            additionalProperties: true
            type: object
      security:
      - JWT: []
      summary: 获取设备通知
      tags:
      - Device
  /device/login:
    post:
      consumes:
      - application/json
      description: 设备登录获取JWT令牌
      parameters:
      - description: 登录信息
        in: body
        name: login
        required: true
        schema:
          type: object
      - description: 设备ID
        in: formData
        name: deviceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回登录令牌和设备信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 错误信息
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 认证失败
          schema:
            additionalProperties: true
            type: object
      security:
      - None: []
      summary: 设备登录
      tags:
      - Device
securityDefinitions:
  BearerAuth:
    description: 输入带有Bearer前缀的用户认证token
    in: header
    name: Authorization
    type: apiKey
  DeviceAuth:
    description: 输入带有Bearer前缀的设备认证token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"